# apps/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESIDA API",
        "description": "REST API untuk sistem informasi manajemen warga perumahan. Mencakup manajemen penghuni, rumah, iuran, pembayaran, pengeluaran, dan laporan keuangan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/houses": {
            "get": {
                "tags": [
                    "Houses"
                ],
                "summary": "Daftar rumah (paginasi)",
                "description": "Daftar rumah dengan paginasi. Mendukung pencarian, filter status hunian, dan sorting.",
                "operationId": "6ec232adfa89735010c2d9475d9f0796",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari nomor rumah",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "occupancy_status",
                        "in": "query",
                        "description": "Filter status hunian",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "dihuni",
                                "tidak_dihuni"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting. Prefix - untuk desc. Kolom: house_number, created_at, occupancy_status.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Data per halaman (max: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data rumah."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Houses"
                ],
                "summary": "Tambah rumah baru",
                "description": "Tambah rumah. Status awal: tidak_dihuni. Nomor rumah harus unik.",
                "operationId": "6aa691dbed6d724cbdfff00b8e869897",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "house_number"
                                ],
                                "properties": {
                                    "house_number": {
                                        "description": "Nomor rumah (unik)",
                                        "type": "string",
                                        "example": "A-01",
                                        "maxLength": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rumah berhasil ditambahkan."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "422": {
                        "description": "Validasi gagal (nomor rumah sudah terdaftar)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/houses/{id}": {
            "get": {
                "tags": [
                    "Houses"
                ],
                "summary": "Detail rumah dengan penghuni aktif",
                "description": "Detail rumah termasuk info penghuni aktif (jika ada).",
                "operationId": "51edc9f63fe720ffa5b79d28b74a8398",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rumah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data rumah berhasil diambil."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Rumah tidak ditemukan."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Houses"
                ],
                "summary": "Update data rumah",
                "description": "Update data rumah. Nomor rumah unik (ignore current).",
                "operationId": "c64a5ca392c2350ec651d8fb5f383135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rumah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "house_number"
                                ],
                                "properties": {
                                    "house_number": {
                                        "type": "string",
                                        "example": "A-01",
                                        "maxLength": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data rumah berhasil diperbarui."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Rumah tidak ditemukan."
                    },
                    "422": {
                        "description": "Validasi gagal."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/houses/{id}/deactivate": {
            "patch": {
                "tags": [
                    "Houses"
                ],
                "summary": "Nonaktifkan rumah (soft delete)",
                "description": "Nonaktifkan rumah. Hanya jika tidak sedang dihuni.",
                "operationId": "416e90937db1883649ee1bf656c5e351",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rumah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rumah berhasil dinonaktifkan."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Rumah tidak ditemukan."
                    },
                    "409": {
                        "description": "Konflik - rumah masih dihuni."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/houses/{id}/occupancy-history": {
            "get": {
                "tags": [
                    "Houses"
                ],
                "summary": "Riwayat penghuni rumah",
                "description": "Seluruh riwayat hunian rumah, diurutkan dari terbaru.",
                "operationId": "0342c0ff441bb5b3039d17073e3c18dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rumah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Riwayat penghuni berhasil diambil."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Rumah tidak ditemukan."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/houses/{id}/payment-history": {
            "get": {
                "tags": [
                    "Houses"
                ],
                "summary": "Riwayat pembayaran rumah",
                "description": "Riwayat pembayaran iuran rumah. Filter berdasarkan jenis iuran dan tahun.",
                "operationId": "2ce3c79c013b28d5857f4beb0c7714f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rumah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dues_type_id",
                        "in": "query",
                        "description": "Filter jenis iuran",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter tahun",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Riwayat pembayaran berhasil diambil."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/houses/{id}/assign-resident": {
            "post": {
                "tags": [
                    "Houses"
                ],
                "summary": "Tempatkan penghuni di rumah",
                "description": "Menempatkan penghuni ke rumah kosong. Validasi: rumah belum dihuni, penghuni belum aktif di rumah lain.",
                "operationId": "83815e8f228577cd4207aa5ef8eda108",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rumah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resident_id"
                                ],
                                "properties": {
                                    "resident_id": {
                                        "description": "ID penghuni",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Penghuni berhasil ditempatkan."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Rumah/penghuni tidak ditemukan."
                    },
                    "409": {
                        "description": "Konflik - rumah sudah dihuni / penghuni sudah aktif."
                    },
                    "422": {
                        "description": "Validasi gagal."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/houses/{id}/reassign-resident": {
            "post": {
                "tags": [
                    "Houses"
                ],
                "summary": "Ganti penghuni rumah",
                "description": "Menutup occupancy lama dan membuat occupancy baru untuk penghuni baru.",
                "operationId": "07334cde1b9c9670c5f9dbc866f60b03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rumah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resident_id"
                                ],
                                "properties": {
                                    "resident_id": {
                                        "description": "ID penghuni baru",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Penghuni berhasil dipindahkan."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Rumah/penghuni tidak ditemukan."
                    },
                    "409": {
                        "description": "Konflik - penghuni baru sudah aktif di rumah lain."
                    },
                    "422": {
                        "description": "Validasi gagal."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Sanctum. Format: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}